create a slice method on the main vtable
