+++ vortex-array/src/arrays/decimal/vtable/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::ops::Range;
+
 use vortex_buffer::Alignment;
 use vortex_buffer::Buffer;
 use vortex_dtype::DType;
@@ -14,6 +16,7 @@ use vortex_scalar::DecimalType;
 
 use crate::ArrayRef;
 use crate::DeserializeMetadata;
+use crate::IntoArray;
 use crate::ProstMetadata;
 use crate::SerializeMetadata;
 use crate::arrays::DecimalArray;
@@ -24,6 +27,7 @@ use crate::vtable;
 use crate::vtable::ArrayVTableExt;
 use crate::vtable::NotSupported;
 use crate::vtable::VTable;
+use crate::vtable::ValidityHelper;
 use crate::vtable::ValidityVTableFromValidityHelper;
 
 mod array;
@@ -148,6 +152,17 @@ impl VTable for DecimalVTable {
         child_idx: usize,
     ) -> VortexResult<Option<ArrayRef>> {
         RULES.evaluate(array, parent, child_idx)
+    }
+
+    fn slice(array: &Self::Array, range: Range<usize>) -> VortexResult<Option<ArrayRef>> {
+        let result = match_each_decimal_value_type!(array.values_type(), |D| {
+            let sliced = array.buffer::<D>().slice(range.clone());
+            let validity = array.validity().clone().slice(range);
+            // SAFETY: Slicing preserves all DecimalArray invariants
+            unsafe { DecimalArray::new_unchecked(sliced, array.decimal_dtype(), validity) }
+                .into_array()
+        });
+        Ok(Some(result))
     }
 }
