+++ vortex-array/src/arrays/extension/vtable/mod.rs
@@ -8,6 +8,8 @@ mod rules;
 mod validity;
 mod visitor;
 
+use std::ops::Range;
+
 use vortex_dtype::DType;
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
@@ -16,6 +18,7 @@ use vortex_error::vortex_ensure;
 
 use crate::ArrayRef;
 use crate::EmptyMetadata;
+use crate::IntoArray;
 use crate::arrays::extension::ExtensionArray;
 use crate::arrays::extension::vtable::rules::PARENT_RULES;
 use crate::buffer::BufferHandle;
@@ -100,6 +103,13 @@ impl VTable for ExtensionVTable {
         child_idx: usize,
     ) -> VortexResult<Option<ArrayRef>> {
         PARENT_RULES.evaluate(array, parent, child_idx)
+    }
+
+    fn slice(array: &Self::Array, range: Range<usize>) -> VortexResult<Option<ArrayRef>> {
+        Ok(Some(
+            ExtensionArray::new(array.ext_dtype().clone(), array.storage().slice(range))
+                .into_array(),
+        ))
     }
 }
