+++ vortex-array/src/arrays/fixed_size_list/vtable/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::ops::Range;
+
 use vortex_dtype::DType;
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
@@ -9,6 +11,7 @@ use vortex_error::vortex_ensure;
 
 use crate::ArrayRef;
 use crate::EmptyMetadata;
+use crate::IntoArray;
 use crate::arrays::FixedSizeListArray;
 use crate::buffer::BufferHandle;
 use crate::serde::ArrayChildren;
@@ -19,6 +22,7 @@ use crate::vtable::ArrayVTable;
 use crate::vtable::ArrayVTableExt;
 use crate::vtable::NotSupported;
 use crate::vtable::VTable;
+use crate::vtable::ValidityHelper;
 use crate::vtable::ValidityVTableFromValidityHelper;
 
 mod array;
@@ -47,6 +51,26 @@ impl VTable for FixedSizeListVTable {
 
     fn id(&self) -> ArrayId {
         ArrayId::new_ref("vortex.fixed_size_list")
+    }
+
+    fn slice(array: &Self::Array, range: Range<usize>) -> VortexResult<Option<ArrayRef>> {
+        let new_len = range.len();
+        let list_size = array.list_size() as usize;
+
+        // SAFETY: Slicing preserves FixedSizeListArray invariants
+        Ok(Some(
+            unsafe {
+                FixedSizeListArray::new_unchecked(
+                    array
+                        .elements()
+                        .slice(range.start * list_size..range.end * list_size),
+                    array.list_size(),
+                    array.validity().slice(range),
+                    new_len,
+                )
+            }
+            .into_array(),
+        ))
     }
 
     fn encoding(_array: &Self::Array) -> ArrayVTable {
