+++ vortex-array/src/arrays/listview/vtable/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::ops::Range;
+
 use vortex_dtype::DType;
 use vortex_dtype::Nullability;
 use vortex_dtype::PType;
@@ -11,6 +13,7 @@ use vortex_error::vortex_ensure;
 
 use crate::ArrayRef;
 use crate::DeserializeMetadata;
+use crate::IntoArray;
 use crate::ProstMetadata;
 use crate::SerializeMetadata;
 use crate::arrays::ListViewArray;
@@ -24,6 +27,7 @@ use crate::vtable::ArrayVTable;
 use crate::vtable::ArrayVTableExt;
 use crate::vtable::NotSupported;
 use crate::vtable::VTable;
+use crate::vtable::ValidityHelper;
 use crate::vtable::ValidityVTableFromValidityHelper;
 
 mod array;
@@ -63,6 +67,22 @@ impl VTable for ListViewVTable {
 
     fn id(&self) -> ArrayId {
         ArrayId::new_ref("vortex.listview")
+    }
+
+    fn slice(array: &Self::Array, range: Range<usize>) -> VortexResult<Option<ArrayRef>> {
+        // SAFETY: Slicing the components of an existing valid array is still valid.
+        Ok(Some(
+            unsafe {
+                ListViewArray::new_unchecked(
+                    array.elements().clone(),
+                    array.offsets().slice(range.clone()),
+                    array.sizes().slice(range.clone()),
+                    array.validity().slice(range),
+                )
+                .with_zero_copy_to_list(array.is_zero_copy_to_list())
+            }
+            .into_array(),
+        ))
     }
 
     fn encoding(_array: &Self::Array) -> ArrayVTable {
