+++ vortex-array/src/arrays/primitive/vtable/mod.rs
@@ -1,8 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use std::ops::Range;
+
 use vortex_dtype::DType;
+use vortex_dtype::NativePType;
 use vortex_dtype::PType;
+use vortex_dtype::match_each_native_ptype;
 use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
@@ -10,6 +14,7 @@ use vortex_error::vortex_ensure;
 
 use crate::ArrayRef;
 use crate::EmptyMetadata;
+use crate::IntoArray;
 use crate::arrays::PrimitiveArray;
 use crate::buffer::BufferHandle;
 use crate::serde::ArrayChildren;
@@ -18,6 +23,7 @@ use crate::vtable;
 use crate::vtable::ArrayVTableExt;
 use crate::vtable::NotSupported;
 use crate::vtable::VTable;
+use crate::vtable::ValidityHelper;
 use crate::vtable::ValidityVTableFromValidityHelper;
 
 mod array;
@@ -144,6 +150,18 @@ impl VTable for PrimitiveVTable {
         child_idx: usize,
     ) -> VortexResult<Option<ArrayRef>> {
         RULES.evaluate(array, parent, child_idx)
+    }
+
+    fn slice(array: &Self::Array, range: Range<usize>) -> VortexResult<Option<ArrayRef>> {
+        let result = match_each_native_ptype!(array.ptype(), |T| {
+            PrimitiveArray::from_buffer_handle(
+                array.buffer_handle().slice_typed::<T>(range.clone()),
+                T::PTYPE,
+                array.validity().slice(range),
+            )
+            .into_array()
+        });
+        Ok(Some(result))
     }
 }
