+++ vortex-array/src/arrays/scalar_fn/vtable/operations.rs
@@ -1,15 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use std::ops::Range;
 use std::sync::Arc;
 
 use vortex_error::VortexExpect;
 use vortex_scalar::Scalar;
 
 use crate::Array;
-use crate::ArrayRef;
-use crate::IntoArray;
 use crate::arrays::scalar_fn::array::ScalarFnArray;
 use crate::arrays::scalar_fn::vtable::ScalarFnVTable;
 use crate::expr::Expression;
