+++ vortex-array/src/arrays/slice/mod.rs
@@ -2,10 +2,8 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 mod array;
-mod execute;
 mod rules;
 mod vtable;
 
 pub use array::*;
-pub use execute::slice_canonical;
 pub use vtable::*;
